Added undo/redo functionality to Settings Window and all HAPI_GUI fields.

+++ Assets/HAPI/Scripts/HAPI_AssetOTLUndoInfo.cs
@@ -28,5 +28,5 @@ public class HAPI_AssetOTLUndoInfo : ScriptableObject
 	// bake settings
 	public float bakeStartTime;
 	public float bakeEndTime;
-	public float bakeSamplesPerSecond;
+	public int bakeSamplesPerSecond;
 }
